Move Exolix exchange-related classes to proper packages
