+++ app/src/main/java/com/tari/android/wallet/ui/screen/exchange/ExchangeFragment.kt
@@ -1,4 +1,4 @@
-package com.tari.android.wallet.ui.screen.home.exchange
+package com.tari.android.wallet.ui.screen.exchange
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -11,7 +11,7 @@ import androidx.fragment.app.viewModels
 import com.tari.android.wallet.data.exolix.CurrencyDto
 import com.tari.android.wallet.ui.common.CommonFragment
 import com.tari.android.wallet.ui.compose.TariDesignSystem
-import com.tari.android.wallet.ui.screen.home.exchange.selectCurrency.SelectCurrencyFragment
+import com.tari.android.wallet.ui.screen.exchange.selectCurrency.SelectCurrencyFragment
 import com.tari.android.wallet.util.extension.composeContent
 import com.tari.android.wallet.util.extension.parcelable
