+++ app/src/main/java/com/tari/android/wallet/ui/screen/exchange/review/ExchangeReviewFragment.kt
@@ -1,4 +1,4 @@
-package com.tari.android.wallet.ui.screen.home.exchange.sendFunds
+package com.tari.android.wallet.ui.screen.exchange.review
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -12,20 +12,20 @@ import com.tari.android.wallet.ui.common.CommonFragment
 import com.tari.android.wallet.ui.compose.TariDesignSystem
 import com.tari.android.wallet.util.extension.composeContent
 
-class SendFundsFragment : CommonFragment<SendFundsViewModel>() {
+class ExchangeReviewFragment : CommonFragment<ExchangeReviewViewModel>() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?) = composeContent {
         val uiState by viewModel.uiState.collectAsState()
 
         TariDesignSystem(viewModel.currentTheme) {
-            SendFundsScreen(
+            ExchangeReviewScreen(
                 uiState = uiState,
                 onBackClick = { viewModel.onBackPressed() },
-                onCopyAmount = { viewModel.onCopyAmount() },
-                onCopyAddress = { viewModel.onCopyAddress() },
-                onCancelTransaction = { viewModel.onBackPressed() },
-                onOpenTxDetails = { viewModel.onOpenTxDetails() },
+                onCopyValueClick = { viewModel.copyValueToClipboard(it) },
+                onConfirmClick = { viewModel.onConfirmClicked() },
                 onRetry = { viewModel.onRetry() },
+                onEmojiIdDetailsClick = { viewModel.onAddressDetailsClicked() },
+                onFeeInfoClick = { viewModel.onFeeInfoClicked() },
             )
         }
     }
@@ -33,14 +33,14 @@ class SendFundsFragment : CommonFragment<SendFundsViewModel>() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        val viewModel: SendFundsViewModel by viewModels()
+        val viewModel: ExchangeReviewViewModel by viewModels()
         bindViewModel(viewModel)
     }
 
     companion object {
         const val ARG_REQUEST = "ARG_REQUEST"
 
-        fun newInstance(request: ExchangeRequestData) = SendFundsFragment().apply {
+        fun newInstance(request: ExchangeRequestData) = ExchangeReviewFragment().apply {
             arguments = Bundle().apply {
                 putParcelable(ARG_REQUEST, request)
             }
