+++ app/src/main/java/com/tari/android/wallet/ui/screen/exchange/status/ExchangeStatusFragment.kt
@@ -1,4 +1,4 @@
-package com.tari.android.wallet.ui.screen.home.exchange.review
+package com.tari.android.wallet.ui.screen.exchange.status
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,25 +7,22 @@ import android.view.ViewGroup
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.fragment.app.viewModels
-import com.tari.android.wallet.data.exolix.ExchangeRequestData
 import com.tari.android.wallet.ui.common.CommonFragment
 import com.tari.android.wallet.ui.compose.TariDesignSystem
 import com.tari.android.wallet.util.extension.composeContent
 
-class ReviewFragment : CommonFragment<ReviewViewModel>() {
+class ExchangeStatusFragment : CommonFragment<ExchangeStatusViewModel>() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?) = composeContent {
         val uiState by viewModel.uiState.collectAsState()
 
         TariDesignSystem(viewModel.currentTheme) {
-            ReviewScreen(
+            ExchangeStatusScreen(
                 uiState = uiState,
                 onBackClick = { viewModel.onBackPressed() },
-                onCopyValueClick = { viewModel.copyValueToClipboard(it) },
-                onConfirmClick = { viewModel.onConfirmClicked() },
-                onRetry = { viewModel.onRetry() },
-                onEmojiIdDetailsClick = { viewModel.onAddressDetailsClicked() },
-                onFeeInfoClick = { viewModel.onFeeInfoClicked() },
+                onTransactionDetailsClick = viewModel::onTransactionDetailsClicked,
+                onCopyClick = viewModel::onCopyClicked,
+                onRetry = viewModel::loadTransaction,
             )
         }
     }
@@ -33,16 +30,16 @@ class ReviewFragment : CommonFragment<ReviewViewModel>() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        val viewModel: ReviewViewModel by viewModels()
+        val viewModel: ExchangeStatusViewModel by viewModels()
         bindViewModel(viewModel)
     }
 
     companion object {
-        const val ARG_REQUEST = "ARG_REQUEST"
+        const val ARG_TRANSACTION_ID = "ARG_TRANSACTION_ID"
 
-        fun newInstance(request: ExchangeRequestData) = ReviewFragment().apply {
+        fun newInstance(transactionId: String) = ExchangeStatusFragment().apply {
             arguments = Bundle().apply {
-                putParcelable(ARG_REQUEST, request)
+                putString(ARG_TRANSACTION_ID, transactionId)
             }
         }
     }
