+++ app/src/main/java/com/tari/android/wallet/ui/screen/exchange/widget/CurrencyItem.kt
@@ -1,4 +1,4 @@
-package com.tari.android.wallet.ui.screen.home.exchange.widget
+package com.tari.android.wallet.ui.screen.exchange.widget
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.clickable
